Add Timetable Fragment
